moving library.h file to src/common/
