+++ src/common/library.h
